+++ tests/atest/transformers/AlignTestCases/test_transformer.py
@@ -28,3 +28,13 @@ class TestAlignTestCases(TransformerAcceptanceTest):
             expected="no_header_col_only_headers.robot",
             config=":only_with_headers=True",
         )
+
+    def test_fixed(self):
+        self.compare(source="test.robot", expected="test_fixed.robot", config=":min_width=30")
+
+    def test_for_fixed(self):
+        self.compare(
+            source="templated_for_loops_and_without.robot",
+            expected="templated_for_loops_and_without_fixed.robot",
+            config=":min_width=25",
+        )