+++ tests/atest/transformers/AlignVariablesSection/test_transformer.py
@@ -35,3 +35,29 @@ class TestAlignVariablesSection(TransformerAcceptanceTest):
 
     def test_multiline_with_blank(self):
         self.compare(source="multiline_with_blank.robot")
+
+    def test_align_variables_skip(self):
+        self.compare(source="tests.robot", expected="tests_skip.robot", config=":skip_types=dict,list")
+
+    def test_align_variables_skip_scalar(self):
+        self.compare(
+            source="align_selected.robot",
+            expected="align_selected_skip.robot",
+            config=":skip_types=scalar",
+        )
+
+    def test_multiline_skip(self):
+        self.compare(source="multiline_skip.robot", config=":skip_types=list,dict")
+
+    def test_fixed_tests(self):
+        self.compare(source="tests.robot", expected="tests_fixed.robot", config=":min_width=30")
+
+    def test_fixed_tests_zero(self):
+        self.compare(source="tests.robot", expected="tests_fixed_one.robot", config=":min_width=1")
+
+    def test_fixed_all_columns(self):
+        self.compare(
+            source="tests.robot",
+            expected="all_columns_fixed.robot",
+            config=":min_width=20:up_to_column=0",
+        )