+++ tests/atest/transformers/DiscardEmptySections/test_transformer.py
@@ -5,13 +5,12 @@ class TestDiscardEmptySections(TransformerAcceptanceTest):
     TRANSFORMER_NAME = "DiscardEmptySections"
 
     def test_removes_empty_sections(self):
-        self.compare(source="removes_empty_sections.robot")
+        self.compare(source="removes_empty_sections.robot", config=":allow_only_comments=False")
 
     def test_removes_empty_sections_except_comments(self):
         self.compare(
             source="removes_empty_sections.robot",
             expected="removes_empty_sections_except_comments.robot",
-            config=":allow_only_comments=True",
         )
 
     def test_remove_selected_empty_node(self):